[BE] Added sites rating raw page.

+++ responser-backend/src/main/java/com/responser/backend/controller/webresource/payload/WebResourceDTO.java
@@ -21,9 +21,9 @@ public class WebResourceDTO {
 
     private WebResourceDTO parent;
 
-    private Float rating;
+    private Double rating;
 
-    private Integer reviewsCount;
+    private Long reviewsCount;
 
     private LocalDateTime creationDate;
